refactor: move models to model directory
