+++ src/cala/gui/nodes/frame_streamer.py
@@ -7,7 +7,7 @@ import numpy as np
 import xarray as xr
 from river.base import Transformer
 
-from cala.streaming.core import Parameters
+from cala.models.params import Parameters
 
 
 @dataclass
