+++ src/cala/streaming/core/__init__.py
@@ -1,9 +1,3 @@
-from .components import Component
-from .parameters import Parameters
 from .stores import ObservableStore
 
-__all__ = [
-    "Parameters",
-    "ObservableStore",
-    "Component",
-]
+__all__ = ["ObservableStore"]
