+++ src/cala/streaming/nodes/init/cold/energy.py
@@ -5,7 +5,7 @@ from scipy.ndimage import gaussian_filter
 from skimage.restoration import estimate_sigma
 from sklearn.feature_extraction.image import PatchExtractor
 
-from cala.streaming.core import Parameters
+from cala.models.params import Parameters
 from cala.streaming.nodes import Node
 from cala.streaming.stores.odl import Residuals
