+++ src/cala/streaming/nodes/init/odl/overlaps.py
@@ -5,7 +5,7 @@ import sparse
 import xarray as xr
 from river.base import SupervisedTransformer
 
-from cala.streaming.core import Parameters
+from cala.models.params import Parameters
 from cala.streaming.stores.common import Footprints
 from cala.streaming.stores.odl import Overlaps
