+++ src/cala/streaming/nodes/init/odl/pixel_stats.py
@@ -4,7 +4,7 @@ from typing import Self
 import xarray as xr
 from river.base import SupervisedTransformer
 
-from cala.streaming.core import Parameters
+from cala.models.params import Parameters
 from cala.streaming.stores.common import Traces
 from cala.streaming.stores.odl import PixelStats
