+++ src/cala/streaming/nodes/iter/pixel_stats.py
@@ -5,7 +5,7 @@ import xarray as xr
 from river.base import SupervisedTransformer
 from sklearn.exceptions import NotFittedError
 
-from cala.streaming.core import Parameters
+from cala.models.params import Parameters
 from cala.streaming.stores.common import Traces
 from cala.streaming.stores.odl import PixelStats
