+++ src/cala/streaming/nodes/preprocess/rigid_stabilization.py
@@ -7,7 +7,7 @@ import xarray as xr
 from river import base
 from skimage.registration import phase_cross_correlation
 
-from cala.streaming.core import Parameters
+from cala.models.params import Parameters
 
 
 @dataclass
