+++ tests/fixtures/mini.py
@@ -5,7 +5,7 @@ import pytest
 import sparse
 import xarray as xr
 
-from cala.streaming.core import Component
+from cala.models.components import Component
 
 
 @dataclass
