+++ tests/test_streaming_service/unit/test_init/test_residual.py
@@ -5,7 +5,7 @@ import pytest
 import xarray as xr
 
 from cala.gui.plots import Plotter
-from cala.streaming.core import Component
+from cala.models.components import Component
 from cala.streaming.nodes.init.odl.residuals import (
     ResidualInitializer,
     ResidualInitializerParams,
