+++ tests/test_streaming_service/unit/test_iter/test_footprints.py
@@ -4,7 +4,7 @@ import xarray as xr
 from scipy.ndimage import binary_dilation, binary_erosion
 
 from cala.gui.plots import Plotter
-from cala.streaming.core import Component
+from cala.models.components import Component
 from cala.streaming.nodes.init.common import (
     TracesInitializer,
     TracesInitializerParams,
