+++ tests/test_streaming_service/unit/test_stores/test_common.py
@@ -4,7 +4,8 @@ import numpy as np
 import pytest
 import xarray as xr
 
-from cala.streaming.core import Component, ObservableStore
+from cala.models.components import Component
+from cala.streaming.core import ObservableStore
 from cala.streaming.stores.common import FootprintStore, TraceStore
