+++ tests/test_streaming_service/unit/test_stores/test_odl.py
@@ -3,7 +3,8 @@ import pytest
 import sparse
 import xarray as xr
 
-from cala.streaming.core import Component, ObservableStore
+from cala.models.components import Component
+from cala.streaming.core import ObservableStore
 from cala.streaming.stores.odl import (
     ComponentStatStore,
     OverlapStore,
